requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.18
    # via langchain-community
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anthropic==0.51.0
    # via langchain-anthropic
anyio==4.9.0
    # via
    #   anthropic
    #   httpx
    #   openai
attrs==25.3.0
    # via aiohttp
cachetools==5.5.2
    # via google-auth
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.4.2
    # via requests
clean-text==0.6.0
    # via -r requirements.in
cryptography==45.0.2
    # via pyjwt
dataclasses-json==0.6.7
    # via langchain-community
deprecated==1.2.18
    # via pygithub
directory-tree==1.0.0
    # via -r requirements.in
distro==1.9.0
    # via
    #   anthropic
    #   openai
emoji==1.7.0
    # via clean-text
faiss-cpu==1.11.0
    # via -r requirements.in
filetype==1.2.0
    # via langchain-google-genai
frozenlist==1.6.0
    # via
    #   aiohttp
    #   aiosignal
ftfy==6.3.1
    # via clean-text
google-ai-generativelanguage==0.6.18
    # via langchain-google-genai
google-api-core==2.24.2
    # via google-ai-generativelanguage
google-auth==2.40.2
    # via
    #   google-ai-generativelanguage
    #   google-api-core
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.71.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.0
    # via google-api-core
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   anthropic
    #   langgraph-sdk
    #   langsmith
    #   openai
httpx-sse==0.4.0
    # via langchain-community
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jiter==0.10.0
    # via
    #   anthropic
    #   openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.25
    # via langchain-community
langchain-anthropic==0.3.13
    # via -r requirements.in
langchain-community==0.3.24
    # via -r requirements.in
langchain-core==0.3.60
    # via
    #   langchain
    #   langchain-anthropic
    #   langchain-community
    #   langchain-google-genai
    #   langchain-openai
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-google-genai==2.1.4
    # via -r requirements.in
langchain-openai==0.3.17
    # via -r requirements.in
langchain-text-splitters==0.3.8
    # via langchain
langgraph==0.4.5
    # via -r requirements.in
langgraph-checkpoint==2.0.26
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==0.1.8
    # via langgraph
langgraph-sdk==0.1.70
    # via langgraph
langsmith==0.3.42
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
marshmallow==3.26.1
    # via dataclasses-json
msgspec==0.19.0
    # via -r requirements.in
multidict==6.4.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
numpy==2.2.6
    # via
    #   faiss-cpu
    #   langchain-community
openai==1.81.0
    # via langchain-openai
orjson==3.10.18
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.9.1
    # via langgraph-checkpoint
packaging==24.2
    # via
    #   faiss-cpu
    #   langchain-core
    #   langsmith
    #   marshmallow
pendulum==3.1.0
    # via -r requirements.in
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.4
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.11.4
    # via
    #   -r requirements.in
    #   anthropic
    #   langchain
    #   langchain-anthropic
    #   langchain-core
    #   langchain-google-genai
    #   langgraph
    #   langsmith
    #   openai
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.9.1
    # via langchain-community
pygithub==2.6.1
    # via -r requirements.in
pyjwt==2.10.1
    # via pygithub
pynacl==1.5.0
    # via pygithub
python-dateutil==2.9.0.post0
    # via pendulum
python-dotenv==1.1.0
    # via
    #   -r requirements.in
    #   pydantic-settings
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
regex==2024.11.6
    # via tiktoken
requests==2.32.3
    # via
    #   google-api-core
    #   langchain
    #   langchain-community
    #   langsmith
    #   pygithub
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
    #   openai
sqlalchemy==2.0.41
    # via
    #   langchain
    #   langchain-community
tenacity==9.1.2
    # via
    #   langchain-community
    #   langchain-core
tiktoken==0.9.0
    # via
    #   -r requirements.in
    #   langchain-openai
tqdm==4.67.1
    # via openai
typing-extensions==4.13.2
    # via
    #   anthropic
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   pygithub
    #   sqlalchemy
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pendulum
urllib3==2.4.0
    # via
    #   pygithub
    #   requests
wcwidth==0.2.13
    # via ftfy
wrapt==1.17.2
    # via deprecated
xxhash==3.5.0
    # via langgraph
yarl==1.20.0
    # via aiohttp
zstandard==0.23.0
    # via langsmith
